fixing build warnings - adding xml docs
